+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] ##
 - Added adoption of SPDX-tags for the distributed manual pages
 - Added better detection of server Wallops
+- Added command-line option `-S`
 - Added extensive use of the `CPPFLAGS` make macro
 - Added handling of empty user@host combinations in Wallops messages
 - Added protection for bad filenames for IRC logs